# Conflicts:
#	src/main/java/com/example/project_bigbangk/repository/RootRepository.java
#	src/main/java/com/example/project_bigbangk/service/OrderMatchingService.java

+++ src/main/java/com/example/project_bigbangk/model/Orders/AbstractOrder.java
@@ -62,7 +62,7 @@ public abstract class AbstractOrder {
         return assetAmount;
     }
 
-    public void setAssetAmount(int assetAmount) {
+    public void setAssetAmount(double assetAmount) {
         this.assetAmount = assetAmount;
     }
